requirements.txt
jupyter
groq
python-dotenv
langchain
langchain-groq
fastapi
uvicorn

